strlen int->size_t

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: anorman <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 09:36:44 by anorman           #+#    #+#             */
-/*   Updated: 2019/05/21 09:36:48 by anorman          ###   ########.fr       */
+/*   Updated: 2019/05/21 11:07:01 by anorman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int		count;
+	size_t	count;
 
 	count = 0;
 	while (str[count] != '\0')
